Patch up BurnerProvider's subprovider
There are 2 Subproviders with an outdated configuration:

* VmSubprovider is created with 'homestead' hardfork configuration.
This seems to make it unable to call some view functions that depend on
newer opcodes. GSN is targeting 'istanbul'.

Skipping VmSubpovider until it is updated should not affect
BurnerProvider.

* HookedWalletSubprovider uses an old 'sigUtil.signTypedData' method. It
has since been updated and exposed as 'signTypedData_v4'. GSN relies on
v4 version of the method to sign transactions as per EIP-712.